Change user choice to an opt-out option
Because the config setting is only stored if the user visits and
saves the form, a default "checked" box will still show up as false
if the user has not saved the config form. So, instead of checking
if an opt-in box was is true, check if an opt out box is false. This
way it will also return false if the user never visits the config
form.