Add styling for nav button text
to keep the bottons the same size and prevent the text from over-
flowing. There doesn't seem to be a good way to have the text
wrap to a second line AND use the elipsis function. That is
because the text-overflow is only triggered when overflow: hidden
AND white-space: nowrap are both set. The nowrap would prevent the
title from going to a second line, which seems like a poor design.
So, going to have to come up with another way to get the elipsis
but pushing these changes for now.